refactor: rewrite AccountAutocomplete and AssignDialog components with TypeScript
