+++ client/src/components/Admin/AccountAutocomplete.tsx
@@ -1,15 +1,30 @@
 import { Autocomplete, TextField } from "@mui/material";
 import { useAccounts } from "../../hooks/useAccounts";
 
+interface Account {
+  id: string | number;
+  firstName: string;
+  lastName: string;
+  email: string;
+}
+interface AccountAutocompleteProps {
+  accountId: string | number | null;
+  setAccount?: (account: Account | null) => void;
+  setAccountId?: (id: string | number | null) => void;
+  label?: string;
+  name?: string
+}
+
 export default function AccountAutocomplete({
   accountId,
   setAccount,
   setAccountId,
   label,
-}) {
+  name
+}: AccountAutocompleteProps) {
   const { data: accounts } = useAccounts();
 
-  const handleChange = (event, value) => {
+  const handleChange = (_event: unknown,value: Account | null) => {
     setAccount && setAccount(value);
     setAccountId && setAccountId(value ? value.id : null);
   };
@@ -21,7 +36,7 @@ export default function AccountAutocomplete({
           value={accounts.find((acct) => acct.id === accountId) || null}
           onChange={handleChange}
           options={accounts}
-          getOptionLabel={(option) =>
+          getOptionLabel={(option: Account) =>
             `${option.lastName}, ${option.firstName} (${option.email})`
           }
           style={{ width: "100%" }}
