+++ client/src/components/Admin/AssignDialog.tsx
@@ -1,0 +1,67 @@
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogProps,
+  DialogTitle,
+} from "@mui/material";
+import PropTypes from "prop-types";
+import { useState } from "react";
+import AccountAutocomplete from "./AccountAutocomplete";
+
+interface AssignDialogProps extends Omit<DialogProps, "onClose"> {
+  open: boolean;
+  onClose: (value: string | number | false) => void;
+}
+
+function AssignDialog(props: AssignDialogProps) {
+  const { onClose, open, ...other } = props;
+  const [accountId, setAccountId] = useState<string | number | null>(null);
+
+  const handleCancel = () => {
+    onClose(false);
+  };
+
+  const handleAssign = () => {
+    if (accountId !== null) {
+      onClose(accountId);
+    }
+
+    return (
+      <Dialog
+        onClose={handleCancel}
+        disableEscapeKeyDown
+        fullWidth
+        maxWidth="sm"
+        aria-labelledby="confirmation-dialog-title"
+        open={open}
+        {...other}
+      >
+        <DialogTitle id="confirmation-dialog-title">Assign to:</DialogTitle>
+        <DialogContent dividers>
+          <AccountAutocomplete
+            name="assignedLoginId"
+            accountId={accountId}
+            setAccountId={(assignedAccountId) =>
+              setAccountId(assignedAccountId)
+            }
+          />
+        </DialogContent>
+        <DialogActions>
+          <Button variant="outlined" autoFocus onClick={handleCancel}>
+            Cancel
+          </Button>
+          <Button
+            variant="contained"
+            onClick={handleAssign}
+            disabled={!accountId}
+          >
+            Assign
+          </Button>
+        </DialogActions>
+      </Dialog>
+    );
+  };
+}
+export default AssignDialog;
